chore: move `nginx-kind-manifests.ts` to `nginx` package
