+++ core/src/plugins/kubernetes/nginx/nginx-kind-manifests.ts
